Add exception classes; add docs.
