+++ firmware-management/updated/updated-client/CMakeLists.txt
@@ -13,4 +13,3 @@ target_link_libraries(updated-client common_compile_warnings)
 target_link_libraries(updated-client updated-rpc)
 
 install(TARGETS updated-client DESTINATION ${CMAKE_INSTALL_BINDIR})
-
