+++ firmware-management/updated/updated-client/source/rpc/Client.h
@@ -15,10 +15,26 @@
 namespace updated {
 namespace rpc {
 
+/**
+ * A client for the UpdateD RPC mechanism.
+ */
 class Client
 {
 public:
+    /**
+     * Create an updated::rpc::Client. 
+     *
+     * Creating a Client object. This includes setting up a channel through
+     * which RPCs can be made by calling the object's member functions.
+     */
     Client();
+
+    /**
+     * Get the update HEADER file for the latest successful firmware update.
+     *
+     * UpdateD treats HEADER files as opaque blobs that are associated with
+     * each update transaction (something like update transaction handles).
+     */
     std::string GetUpdateHeader();
 
 private:
