+++ firmware-management/updated/updated-rpc/source/config.cpp
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
-#include <updated-rpc/config.h>
+#include "updated-rpc/config.h"
 
 // This file is just to make sure that config.h compiles and is checked by
 // static analyzers.
