+++ firmware-management/updated/updated/source/rpc/Server.h
@@ -13,11 +13,34 @@
 namespace updated {
 namespace rpc {
 
+/**
+ * Class for UpdateD RPC servers.
+ *
+ * A server is started when a Server object is created and is shut down when
+ * the object is destroyed.
+ *
+ * The server listens for RPC requests and services them using an
+ * updated::rpc::ServiceImpl object.
+ */
 class Server final
 {
 public:
+    /**
+     * Create an updated::rpc::Server.
+     *
+     * This includes setting up a socket for listening and starting a new
+     * thread (or threads) for servicing RPC requests.
+     */
     Server();
+
+    /**
+     * Shut down the RPC server and wait for in-progress RPCs to finish.
+     */
     ~Server() noexcept;
+
+    /**
+     * Asynchronously send a shutdown message to the RPC server.
+     */
     void shut_down();
 
 private:
