+++ firmware-management/updated/updated/source/rpc/ServiceImpl.h
@@ -11,6 +11,9 @@
 namespace updated {
 namespace rpc {
 
+/**
+ * Implements the UpdateD RPC service. Methods should only be called from gRPC.
+ */
 class ServiceImpl final
     : public UpdateDService::Service
 {
@@ -18,12 +21,23 @@ public:
     ServiceImpl() = default;
 
 private:
+    /**
+     * Implement the GetUpdateHeader RPC.
+     *
+     * This RPC returns the contents of the update HEADER file for the most
+     * recent successful firmware update transaction.
+     */
     grpc::Status GetUpdateHeader(
         grpc::ServerContext*,
         const Empty*,
         GetUpdateHeaderResponse* response
     ) override;
 
+    /**
+     * Implement the StartUpdate RPC.
+     *
+     * This RPC asks UpdateD to start a new update transaction.
+     */
     grpc::Status StartUpdate(
         grpc::ServerContext*,
         const StartUpdateRequest* request,
